+++ src/paywall-previews/Preview2/Preview2.tsx
@@ -9,7 +9,7 @@ import colors from '../../theme/colors';
 import Branding from '../types/branding';
 
 // Components
-import TextEditor from '../shared/TextEditor';
+import PreviewText from '../shared/PreviewText';
 import { StyledPaywallDescription, BuyButton, FooterLink } from '../shared/PreviewComponents';
 import {
   StyledPreviewBox,
@@ -86,7 +86,7 @@ const Preview2 = ({
       height={height}
       width={width}
     >
-      <StyledImageHolder backgroundImage={image} />
+      <StyledImageHolder backgroundImage={image} onClick={handleOpenModal} />
       <AssetDetails>
         <StyledPaywallDescription color={buttonBgColor} hasProtectedByLabel={hasProtectedByLabel}>
           <PaywallDescriptionSpan>
@@ -94,8 +94,8 @@ const Preview2 = ({
             {premiumContentLabel}
           </PaywallDescriptionSpan>
         </StyledPaywallDescription>
-        <TextEditor value={previewTitle} readOnly padding="0 0 0.5rem 0" />
-        <TextEditor value={previewDescription} readOnly />
+        <PreviewText value={previewTitle} padding="0 0 0.5rem 0" />
+        <PreviewText value={previewDescription} />
         <BuyButtonWrapper>
           <BuyButton
             buttonBgColor={buttonBgColor}